支持通过C++ API查看特定设备按键能力支持情况

+++ service/device_manager/src/input_device_manager.cpp
@@ -66,7 +66,7 @@ std::map<int32_t, bool> InputDeviceManager::GetKeystrokeAbility(int32_t deviceId
         return keystrokeAbility;
     }
     for (const auto& item : keyCodes) {
-        auto sysKeyCode = static_cast<int32_t>(InputTransformationKeyValue(item));
+        auto sysKeyCode = InputTransformationKeyValue(item);
         bool ret = libinput_device_has_key(iter->second, sysKeyCode) == 1 ? true : false;
         keystrokeAbility[item] = ret;
     }
