+++ service/event_handler/src/key_event_value_transformation.cpp
@@ -27,7 +27,7 @@ constexpr OHOS::HiviewDFX::HiLogLabel LABEL = { LOG_CORE, MMI_LOG_DOMAIN, "KeyEv
 constexpr int32_t INVALID_KEY_CODE = -1;
 } // namespace
 
-const std::multimap<int16_t, KeyEventValueTransformations> MAP_KEY_EVENT_VALUE_TRANSFORMATION = {
+const std::multimap<int32_t, KeyEventValueTransformations> MAP_KEY_EVENT_VALUE_TRANSFORMATION = {
     {11, {"KEY_0", 11, 2000, HOS_KEY_0}},
     {2, {"KEY_1", 2, 2001, HOS_KEY_1}},
     {3, {"KEY_2", 3, 2002, HOS_KEY_2}},
@@ -444,11 +444,11 @@ const std::multimap<int16_t, KeyEventValueTransformations> MAP_KEY_EVENT_VALUE_T
     {412, {"KEY_PREVIOUS", 412, 2631, HOS_KEY_PREVIOUS}},
 };
 
-KeyEventValueTransformations KeyValueTransformationInput(int16_t keyValueOfInput)
+KeyEventValueTransformations KeyValueTransformationInput(int32_t keyValueOfInput)
 {
     auto it = MAP_KEY_EVENT_VALUE_TRANSFORMATION.find(keyValueOfInput);
     if (it == MAP_KEY_EVENT_VALUE_TRANSFORMATION.end()) {
-        constexpr int16_t UNKNOWN_KEY_BASE = 10000;
+        constexpr int32_t UNKNOWN_KEY_BASE = 10000;
         KeyEventValueTransformations unknownKey = {
             "UNKNOWN_KEY", keyValueOfInput, UNKNOWN_KEY_BASE + keyValueOfInput, HOS_UNKNOWN_KEY_BASE
         };
@@ -459,10 +459,10 @@ KeyEventValueTransformations KeyValueTransformationInput(int16_t keyValueOfInput
     return it->second;
 }
 
-int16_t InputTransformationKeyValue(int32_t keyCode)
+int32_t InputTransformationKeyValue(int32_t keyCode)
 {
     for (const auto &item : MAP_KEY_EVENT_VALUE_TRANSFORMATION) {
-        if (item.second.keyValueOfSys == static_cast<int16_t>(keyCode)) {
+        if (item.second.keyValueOfSys == keyCode) {
             return item.first;
         }
     }
